Refactored Setttings
