+++ tests/foundry/client-esm/client.test-d.ts
@@ -1,0 +1,2 @@
+// @ts-expect-error The `types` file is not actually exported as an object into the foundry namespace
+foundry.types
